refactor(api): move saml controller
to saml directory
