+++ api/src/identity-access-management/application/routes.js
@@ -1,6 +1,6 @@
 import { oidcProviderAdminRoutes } from './oidc-provider/oidc-provider.admin.route.js';
 import { oidcProviderRoutes } from './oidc-provider/oidc-provider.route.js';
-import { samlRoutes } from './saml.route.js';
+import { samlRoutes } from './saml/saml.route.js';
 import { tokenRoutes } from './token.route.js';
 import { userAdminRoutes } from './user/user.admin.route.js';
 import { userRoutes } from './user/user.route.js';
