+++ api/src/identity-access-management/application/saml/saml.controller.js
@@ -1,8 +1,8 @@
-import { config } from '../../../lib/config.js';
-import { tokenService } from '../../shared/domain/services/token-service.js';
-import { logger } from '../../shared/infrastructure/utils/logger.js';
-import { usecases } from '../domain/usecases/index.js';
-import * as saml from '../infrastructure/saml.js';
+import { config } from '../../../../lib/config.js';
+import { tokenService } from '../../../shared/domain/services/token-service.js';
+import { logger } from '../../../shared/infrastructure/utils/logger.js';
+import { usecases } from '../../domain/usecases/index.js';
+import * as saml from '../../infrastructure/saml.js';
 
 const metadata = function (request, h) {
   return h.response(saml.getServiceProviderMetadata()).type('application/xml');
